Stop expecting replies for commands.
According to ZigBee Cluster Library SpecificationRevision 6 Draft Version 1.0, Section 2.5.12.2 "When Generated", zigbee is not required to issue an explicit reply. If the initiating request is expecting a Default Response then any message sent with the same Transaction sequence number can act as a reply. The implementation to track what counts as a stand-in for a Default Response would be nontrivial.

This change fixes the behavior for the most common case of sending commands by setting expect_reply=False for all commands. This will most likely break other use cases where a command requires a response (ie. the Get Alarm request). In the longer term, the detection of the Default Response stand-in needs to be implemented or the commands in cluster definitions need to be annotated to specify if they require a reply.

This fix allows zigby to control https://www.homedepot.com/p/Hampton-Bay-Universal-Wink-Enabled-White-Ceiling-Fan-Premier-Remote-Control-99432/206591100.