Update README.md with a more descriptive description
Update README.md with a more descriptive description, something that might be useful to those new to Zigbee and/or new to Home Assistant. Hopefully, this description simplified to use more layman terms yet with more detailed and links might help make new developers make an informed decision if this project is something that they want to use or help out with.