fix: avoid waiting on checkin response response
Avoid waiting on the Default_Response for a checkin_response request. There is no use for
the reply data.

Ensure we set disable_default_response on the frames whenever we are not expecting a reply.
It does not matter if this is a client or server request,
what matters is if we are going to be waiting for a
reply.